requirements.txt
python-binance==1.0.19
typer==0.9.0
python-dotenv==1.0.1